* thread.c: Fixed implicit conversion error with Apple clang-800.0.31
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@55619 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ thread.c
@@ -1085,7 +1085,7 @@ getclockofday(struct timeval *tp)
 
     if (clock_gettime(CLOCK_MONOTONIC, &ts) == 0) {
 	tp->tv_sec = ts.tv_sec;
-	tp->tv_usec = ts.tv_nsec / 1000;
+	tp->tv_usec = (int)(ts.tv_nsec / 1000);
     }
     else
 #endif
